Separate features into Features class
