+++ src/Loader.php
@@ -7,9 +7,9 @@
 namespace Automattic\WooCommerce\Admin;
 
 use \_WP_Dependency;
-use Automattic\WooCommerce\Admin\Features\Onboarding;
 use Automattic\WooCommerce\Admin\API\Reports\Orders\DataStore as OrdersDataStore;
 use Automattic\WooCommerce\Admin\API\Plugins;
+use Automattic\WooCommerce\Admin\Features\Features;
 use Automattic\WooCommerce\Admin\Features\Navigation\Screen;
 use WC_Marketplace_Suggestions;
 
@@ -58,12 +58,8 @@ class Loader {
 	 * Hooks added here should be removed in `wc_admin_initialize` via the feature plugin.
 	 */
 	public function __construct() {
+		Features::get_instance();
 		add_action( 'init', array( __CLASS__, 'define_tables' ) );
-		// Load feature before WooCommerce update hooks.
-		add_action( 'init', array( __CLASS__, 'load_features' ), 4 );
-		add_filter( 'woocommerce_get_sections_advanced', array( __CLASS__, 'add_features_section' ) );
-		add_filter( 'woocommerce_get_settings_advanced', array( __CLASS__, 'add_features_settings' ), 10, 2 );
-		add_filter( 'woocommerce_get_settings_advanced', array( __CLASS__, 'maybe_load_beta_features_modal' ), 10, 2 );
 		add_action( 'admin_enqueue_scripts', array( __CLASS__, 'register_scripts' ) );
 		add_action( 'admin_enqueue_scripts', array( __CLASS__, 'inject_wc_settings_dependencies' ), 14 );
 		add_action( 'admin_enqueue_scripts', array( __CLASS__, 'load_scripts' ), 15 );
@@ -140,9 +136,11 @@ class Loader {
 	 * Gets a build configured array of enabled WooCommerce Admin features/sections.
 	 *
 	 * @return array Enabled Woocommerce Admin features/sections.
+	 *
+	 * @deprecated since 1.9.0, use Features::get_features()
 	 */
 	public static function get_features() {
-		return apply_filters( 'woocommerce_admin_features', array() );
+		return Features::get_features();
 	}
 
 	/**
@@ -176,10 +174,11 @@ class Loader {
 	 *
 	 * @param  string $feature Feature slug.
 	 * @return bool Returns true if the feature is enabled.
+	 *
+	 * @deprecated since 1.9.0, use Features::is_enabled( $feature )
 	 */
 	public static function is_feature_enabled( $feature ) {
-		$features = self::get_features();
-		return in_array( $feature, $features, true );
+		return Features::is_enabled( $feature );
 	}
 
 	/**
@@ -190,7 +189,7 @@ class Loader {
 	 */
 	public static function should_use_minified_js_file( $script_debug ) {
 		// minified files are only shipped in non-core versions of wc-admin, return false if minified files are not available.
-		if ( ! self::is_feature_enabled( 'minified-js' ) ) {
+		if ( ! Features::is_enabled( 'minified-js' ) ) {
 			return false;
 		}
 
@@ -238,119 +237,6 @@ class Loader {
 	 */
 	private static function get_path( $ext ) {
 		return ( 'css' === $ext ) ? WC_ADMIN_DIST_CSS_FOLDER : WC_ADMIN_DIST_JS_FOLDER;
-	}
-
-	/**
-	 * Class loader for enabled WooCommerce Admin features/sections.
-	 */
-	public static function load_features() {
-		$features = self::get_features();
-		foreach ( $features as $feature ) {
-			$feature       = str_replace( '-', '', ucwords( strtolower( $feature ), '-' ) );
-			$feature_class = 'Automattic\\WooCommerce\\Admin\\Features\\' . $feature;
-
-			// Handle features contained in subdirectory.
-			if ( ! class_exists( $feature_class ) && class_exists( $feature_class . '\\Init' ) ) {
-				$feature_class = $feature_class . '\\Init';
-			}
-
-			if ( class_exists( $feature_class ) ) {
-				new $feature_class();
-			}
-		}
-	}
-
-	/**
-	 * Adds the Features section to the advanced tab of WooCommerce Settings
-	 *
-	 * @param array $sections Sections.
-	 * @return array
-	 */
-	public static function add_features_section( $sections ) {
-		$features = apply_filters(
-			'woocommerce_settings_features',
-			array()
-		);
-
-		if ( empty( $features ) ) {
-			return $sections;
-		}
-
-		$sections['features'] = __( 'Features', 'woocommerce-admin' );
-		return $sections;
-	}
-
-	/**
-	 * Conditionally loads the beta features tracking modal.
-	 *
-	 * @param array $settings Settings.
-	 * @return array
-	 */
-	public static function maybe_load_beta_features_modal( $settings ) {
-		$tracking_enabled = get_option( 'woocommerce_allow_tracking', 'no' );
-
-		if ( 'yes' === $tracking_enabled ) {
-			return $settings;
-		}
-
-		$rtl = is_rtl() ? '.rtl' : '';
-
-		wp_enqueue_style(
-			'wc-admin-beta-features-tracking-modal',
-			self::get_url( "beta-features-tracking-modal/style{$rtl}", 'css' ),
-			array( 'wp-components' ),
-			self::get_file_version( 'css' )
-		);
-
-		wp_enqueue_script(
-			'wc-admin-beta-features-tracking-modal',
-			self::get_url( 'wp-admin-scripts/beta-features-tracking-modal', 'js' ),
-			array( 'wp-i18n', 'wp-element', WC_ADMIN_APP ),
-			self::get_file_version( 'js' ),
-			true
-		);
-
-		return $settings;
-	}
-
-	/**
-	 * Adds the Features settings.
-	 *
-	 * @param array  $settings Settings.
-	 * @param string $current_section Current section slug.
-	 * @return array
-	 */
-	public static function add_features_settings( $settings, $current_section ) {
-		if ( 'features' !== $current_section ) {
-			return $settings;
-		}
-
-		$features = apply_filters(
-			'woocommerce_settings_features',
-			array()
-		);
-
-		if ( empty( $features ) ) {
-			return $settings;
-		}
-
-		return array_merge(
-			array(
-				array(
-					'title' => __( 'Features', 'woocommerce-admin' ),
-					'type'  => 'title',
-					'desc'  => __( 'Test new features to improve the store management experience. These features might be included in future versions of WooCommerce. <a href="https://href.li/?https://woocommerce.com/usage-tracking/">Requires usage tracking.</a>', 'woocommerce-admin' ),
-					'id'    => 'features_options',
-				),
-			),
-			$features,
-			array(
-				array(
-					'type' => 'sectionend',
-					'id'   => 'features_options',
-				),
-			)
-		);
 	}
 
 	/**
@@ -826,13 +712,6 @@ class Loader {
 		// Grab translation strings from Webpack-generated chunks.
 		add_filter( 'load_script_translation_file', array( __CLASS__, 'load_script_translation_file' ), 10, 3 );
 
-		$features         = self::get_features();
-		$enabled_features = array();
-		foreach ( $features as $key ) {
-			$enabled_features[ $key ] = self::is_feature_enabled( $key );
-		}
-		wp_add_inline_script( WC_ADMIN_APP, 'window.wcAdminFeatures = ' . wp_json_encode( $enabled_features ), 'before' );
-
 		wp_enqueue_script( WC_ADMIN_APP );
 		wp_enqueue_style( WC_ADMIN_APP );
 		wp_enqueue_style( 'wc-material-icons' );
@@ -1040,11 +919,6 @@ class Loader {
 
 		if ( self::is_admin_page() && $is_loading ) {
 			$classes[] = 'woocommerce-admin-is-loading';
-		}
-
-		$features = self::get_features();
-		foreach ( $features as $feature_key ) {
-			$classes[] = sanitize_html_class( 'woocommerce-feature-enabled-' . $feature_key );
 		}
 
 		$admin_body_class = implode( ' ', array_unique( $classes ) );
